Revert "add:Added the Home Page"
