+++ package.json
@@ -15,14 +15,12 @@
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
     "firebase": "^10.0.0",
-    "framer-motion": "^10.13.0",
+    "framer-motion": "^10.12.21",
     "lottie-react": "^2.4.0",
-    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
-    "react-router-dom": "^6.14.2",
-    "styled-components": "^6.0.5"
+    "react-router-dom": "^6.14.2"
   },
   "devDependencies": {
     "@types/react": "^18.2.14",
