--- src/TempData/Services.jsx
@@ -1,17 +1,0 @@
-export default [
-  {
-    id: 1,
-    title: 'Wash & Iron',
-    image: 'Wash.svg',
-  },
-  {
-    id: 2,
-    title: 'Power clean',
-    image: 'PowerClean.svg',
-  },
-  {
-    id: 3,
-    title: 'Dry Clean',
-    image: 'DryClean.svg',
-  },
-];
