--- src/TempData/Workings.jsx
@@ -1,20 +1,0 @@
-export default [
-  {
-    id: 1,
-    image: 'Calendar.svg',
-    title: 'Schedule Pickup Date',
-    desc: 'Easily choose collection & delivery times at you convenience',
-  },
-  {
-    id: 2,
-    image: 'Delivery.svg',
-    title: 'We Pick Up Your Clothes',
-    desc: 'Our driver brings your times to our cleaning partners, where we take utmost care to ensure great results ',
-  },
-  {
-    id: 3,
-    image: 'Tshirt.svg',
-    title: 'Get Clean Laundry',
-    desc: 'Your clothes are back with you 24 hours later - all clean and ready for action',
-  },
-];
